requirements.txt
flask
torch
torchvision
transformers
pillow
opencv-python
numpy